+++ src/main/java/com/ultikits/plugins/kits/gui/KitBrowserGui.java
@@ -212,6 +212,9 @@ public class KitBrowserGui extends Gui {
             case INSUFFICIENT_FUNDS:
                 player.sendMessage(ChatColor.RED + plugin.i18n("余额不足"));
                 break;
+            case PAYMENT_FAILED:
+                player.sendMessage(ChatColor.RED + plugin.i18n("扣款失败，礼包未领取"));
+                break;
             case ALREADY_CLAIMED:
                 player.sendMessage(ChatColor.RED + plugin.i18n("你已经领取过此礼包"));
                 break;